Included changes from 3vcloud repo

+++ code/client/merchant.c
@@ -30,8 +30,8 @@ void GameSrv_TransactItems(GwClient *client, TransactionType type,
     packet.gold_send = gold_send;
     assert(0 <= send_info->item_count && send_info->item_count <= 16);
 
-    packet.n_item_ids_send = send_info->item_count;
-    packet.n_item_quant_send = send_info->item_count;
+    packet.n_item_ids_send = (uint32_t)send_info->item_count;
+    packet.n_item_quant_send = (uint32_t)send_info->item_count;
     for (size_t i = 0; i < send_info->item_count; i++) {
         packet.item_ids_send[i] = send_info->item_ids[i];
         packet.item_quant_send[i] = (int8_t)send_info->item_quants[i];
@@ -40,8 +40,8 @@ void GameSrv_TransactItems(GwClient *client, TransactionType type,
     packet.gold_recv = gold_recv;
     assert(0 <= recv_info->item_count && recv_info->item_count <= 16);
 
-    packet.n_item_ids_recv = recv_info->item_count;
-    packet.n_item_quant_recv = recv_info->item_count;
+    packet.n_item_ids_recv = (uint32_t)recv_info->item_count;
+    packet.n_item_quant_recv = (uint32_t)recv_info->item_count;
     for (size_t i = 0; i < recv_info->item_count; i++) {
         packet.item_ids_recv[i] = recv_info->item_ids[i];
         packet.item_quant_recv[i] = (int8_t)recv_info->item_quants[i];
@@ -74,14 +74,14 @@ void GameSrv_BuyMaterials(GwClient *client, TransactionType type,
     packet.gold_send = gold_send;
     assert(0 <= send_info->item_count && send_info->item_count <= 16);
 
-    packet.n_item_ids_send = send_info->item_count;
+    packet.n_item_ids_send = (uint32_t)send_info->item_count;
     for (size_t i = 0; i < send_info->item_count; i++) {
         packet.item_ids_send[i] = send_info->item_ids[i];
     }
 
     packet.gold_recv = gold_recv;
     assert(0 <= recv_info->item_count && recv_info->item_count <= 16);
-    packet.n_item_ids_recv = recv_info->item_count;
+    packet.n_item_ids_recv = (uint32_t)recv_info->item_count;
     for (size_t i = 0; i < recv_info->item_count; i++) {
         packet.item_ids_recv[i] = recv_info->item_ids[i];
     }
@@ -110,12 +110,12 @@ void GameSrv_RequestQuote(GwClient *client, TransactionType type,
     packet.type = type;
     
     packet.unk_send = 0;
-    packet.n_item_ids_send= send_info->item_count;
+    packet.n_item_ids_send= (uint32_t)send_info->item_count;
     for (size_t i = 0; i < send_info->item_count; i++)
         packet.item_ids_send[i] = send_info->item_ids[i];
 
     packet.unk_recv = 0;
-    packet.n_item_ids_recv= recv_info->item_count;
+    packet.n_item_ids_recv= (uint32_t)recv_info->item_count;
     for (size_t i = 0; i < recv_info->item_count; i++)
         packet.item_ids_recv[i] = recv_info->item_ids[i];
